move models location
